+++ .storybook/preview.tsx
@@ -1,6 +1,8 @@
 import React from "react";
+
 import "../src/styles/constants.css";
-import "../src/styles/colors.module.css";
+import "../src/styles/reset.css";
+import "../src/styles/globals.css";
 import useTheme from "../src/hooks/useTheme";
 import { UserStyleContextProvider } from "../src/context/UserStyleContext";
 
@@ -19,7 +21,15 @@ const withThemeProvider = (Story, context) => {
   useTheme();
   return (
     <UserStyleContextProvider>
-      <Story {...context} />;
+      <link
+        href="https://fonts.googleapis.com/css2?family=Lexend:wght@300&display=swap"
+        rel="stylesheet"
+      />
+      <link
+        href="https://fonts.googleapis.com/css2?family=ABeeZee&display=swap"
+        rel="stylesheet"
+      />
+      <Story {...context} />
     </UserStyleContextProvider>
   );
 };