+++ src/pages/_app.tsx
@@ -2,10 +2,10 @@ import { FC } from "react";
 import type { AppProps } from "next/app";
 import { ApolloProvider } from "@apollo/client";
 
-import { AuthProvider } from "../auth/useAuth";
 import "../styles/constants.css";
 import "../styles/reset.css";
 import "../styles/globals.css";
+import { AuthProvider } from "../context/Auth";
 import useApolloClient from "../browser-lib/graphql/useApolloClient";
 import useTheme from "../hooks/useTheme";
 import { UserStyleContextProvider } from "../context/UserStyleContext";