+++ CHANGE_LOG.md
@@ -1,3 +1,41 @@
+## [0.19.3](https://github.com/oaknational/Samara/compare/v0.19.2...v0.19.3) (2022-05-17)
+
+
+### Bug Fixes
+
+* remove swc darwin package so that node builds on linux ([4087376](https://github.com/oaknational/Samara/commit/4087376d158386918a8b331402d04417baf332e2))
+* swc failed to load error on M1 macs ([2808c9b](https://github.com/oaknational/Samara/commit/2808c9b0a169e58e3547c6541ce450915f3934a8))
+
+## [0.19.2](https://github.com/oaknational/Samara/compare/v0.19.1...v0.19.2) (2022-05-17)
+
+## [0.19.1](https://github.com/oaknational/Samara/compare/v0.19.0...v0.19.1) (2022-05-17)
+
+
+### Bug Fixes
+
+* **global-imports:** global css imports to be at top of _app.tsx ([3f0e4cd](https://github.com/oaknational/Samara/commit/3f0e4cd4ac7b9f188f3b825077b905d72c1d8479))
+* **storybook:** add fonts and global styles to storybook preview ([88aab14](https://github.com/oaknational/Samara/commit/88aab1472bdbd7e2c0f8780dfa6ecd6d6d236242))
+
+# [0.19.0](https://github.com/oaknational/Samara/compare/v0.18.0...v0.19.0) (2022-05-17)
+
+
+### Features
+
+* custom 404 and 500 pages ([eb6ee76](https://github.com/oaknational/Samara/commit/eb6ee760b69f50c62259d1e29bda97a98169b9dc))
+
+# [0.18.0](https://github.com/oaknational/Samara/compare/v0.17.0...v0.18.0) (2022-05-17)
+
+
+### Bug Fixes
+
+* add react-hooks eslint plugin ([8e002b7](https://github.com/oaknational/Samara/commit/8e002b7790e598b5ea87f18943bcf7c78da40b05))
+* add react-hooks eslint plugin ([4635ede](https://github.com/oaknational/Samara/commit/4635edeba1bf37255908cc1205e5f76c984ed1a1))
+
+
+### Features
+
+* **bookmarks:** optimistically add/remove bookmark ([524a874](https://github.com/oaknational/Samara/commit/524a874fd735748a248f06cb43f0f387f397ee9f))
+
 # [0.17.0](https://github.com/oaknational/Samara/compare/v0.16.2...v0.17.0) (2022-05-13)